requirements.txt
streamlit
langchain
PyPdf2
python-dotenv